Category now works!

+++ src/main/java/ch/zli/m223/punchclock/controller/EntryController.java
@@ -1,11 +1,15 @@
 package ch.zli.m223.punchclock.controller;
 
+import ch.zli.m223.punchclock.domain.Category;
 import ch.zli.m223.punchclock.domain.Entry;
+import ch.zli.m223.punchclock.domain.User;
+import ch.zli.m223.punchclock.domain.UserDetailsServiceImpl;
 import ch.zli.m223.punchclock.service.EntryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import javax.ws.rs.BadRequestException;
+import java.security.Principal;
 import java.util.List;
 
 /**
@@ -17,9 +21,11 @@ import java.util.List;
 @RequestMapping("/entries")
 public class EntryController {
     private EntryService entryService;
+    private UserDetailsServiceImpl userDetailsService;
 
-    public EntryController(EntryService entryService) {
+    public EntryController(EntryService entryService, UserDetailsServiceImpl userDetailsService) {
         this.entryService = entryService;
+        this.userDetailsService = userDetailsService;
     }
 
     @GetMapping
@@ -30,14 +36,25 @@ public class EntryController {
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Entry createEntry(@Valid @RequestBody Entry entry) {
+    public Entry createCategory(@RequestBody Entry entry, Principal principal) {
+        User applicationUser = userDetailsService.getUserByUsername(principal.getName());
+        if(!applicationUser.getRole().equals("admin")) throw new BadRequestException();
         return entryService.createEntry(entry);
     }
 
-    //Delete implementieren
-    @DeleteMapping("{id}")
-    @ResponseStatus(HttpStatus.OK)
-    public void deleteEntry(@PathVariable Long id) {
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteEntry(@PathVariable long id, Principal principal) {
+        User applicationUser = userDetailsService.getUserByUsername(principal.getName());
+        if(!applicationUser.getRole().equals("admin")) throw new BadRequestException();
         entryService.deleteEntry(id);
+    }
+
+    @PutMapping("/{id}")
+    @ResponseStatus(HttpStatus.OK)
+    public Entry updateEntry(@RequestBody Entry entry, Principal principal) {
+        User applicationUser = userDetailsService.getUserByUsername(principal.getName());
+        if (!applicationUser.getRole().equals("admin")) throw new BadRequestException();
+        return entryService.updateEntry(entry);
     }
 }
